Convert  example to TypeScript
